requirements.txt
numpy
scipy
scikit-learn
hnswlib
faiss-cpu
python-igraph
leidenalg
hdbscan
PyYAML
joblib
tqdm
metric-learn
